Replace anonymous FluxCCrashLogger with injectable logger
